% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_interaction_qtl.R
\name{map_interaction_qtl}
\alias{map_interaction_qtl}
\title{Response (e)QTL mapping}
\usage{
map_interaction_qtl(input, genotype, geneloc, snploc, anno,
  sample_kernel = NULL, normalization_approach = "qq",
  permutation_approach = "boot", cisdist = 1e+05, checkpoint_dir = NULL,
  debug = F)
}
\arguments{
\item{input}{[genes x samples] matrix of log_2 expression values, e.g. cpm or fpkm}

\item{genotype}{[SNPs x individuals] matrix of genotypes (can have some NAs)}

\item{geneloc}{[genes x 4] data.frame with columns (geneid, chr, left, right). The intersect of geneloc$geneid and rownames(input) will be tested.}

\item{snploc}{[SNPs x 3] data.frame with columns (snpid, chr, pos) where snpid's map to rows of `genotype`}

\item{anno}{[samples x 2] data.frame with columns (individual, condition) where individual must correspond to column names of `genotype`}

\item{sample_kernel}{[samples x samples] covariance matrix learned in suez step 1. If you don't want to control for latent confounders leave this as NULL to use the matrix representing which samples are from the same individual.}

\item{normalization_approach}{One of (qq,log,linear) determining how to normalize within each gene. We recommend qq (quantile normalization to normal) to ensure the assumptions of the test hold.}

\item{permutation_approach}{One of ("none","permute","boot"). We recommend boot (parametric bootstrap) since the permutation test is not really valid for interaction effect testing.}

\item{cisdist}{How far from the gene boundaries to look for SNPs to test}

\item{checkpoint_dir}{An optional directory to store per gene results. If mapping crashes these results will be reused to save time.}

\item{debug}{Controls whether you see stan::optimizing messages and whether errors are suppressed.}
}
\description{
Response (e)QTL mapping
}
