% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_interaction_qtl.R
\name{qqnorm_no_plot}
\alias{qqnorm_no_plot}
\title{qqnorm without plotting}
\usage{
qqnorm_no_plot(x)
}
\arguments{
\item{x}{Numeric vector input}
}
\value{
Quantile normalized version
}
\description{
qqnorm without plotting
}
